GPS return HDOP as double not int

+++ lib/system_utils/gps.h
@@ -136,7 +136,7 @@ int16_t GpsGetLatestGpsAltitude( void );
  *
  * \retval hdop
  */
-int8_t GpsGetLatestGpsHorizontalDilution ( void );
+double GpsGetLatestGpsHorizontalDilution ( void );
 
 
 /*!
